requirements.txt
APScheduler==3.10.1
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==3.2.18
django-cors-headers==3.14.0
django-dotenv==1.4.2
django-environ==0.10.0
django-mssql-backend==2.8.1
djangorestframework==3.14.0
djangorestframework-csv==2.1.1
gunicorn==20.1.0
idna==3.4
mssql-django==1.2
numpy==1.24.2
pandas==1.5.3
psycopg2-binary==2.9.6
pyodbc==4.0.35
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tzdata==2023.3
tzlocal==4.3
unicodecsv==0.14.1
urllib3==1.26.15
whitenoise==6.4.0
